pyproject.toml
[project]
name = "beacon-rag-assistant"
version = "0.1.0"
description = "RAG-powered knowledge assistant with cited, grounded answers — built with ChromaDB, Claude, and Streamlit"
authors = [{ name = "Kirpal Sachdev" }]
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "streamlit>=1.38.0",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
]

[project.scripts]
beacon = "beacon.app:main"
